+++ codegen/compiler/src/Quidditch/Conversion/ConvertDMAToLLVM.h
@@ -1,0 +1,10 @@
+
+#pragma once
+
+#include "mlir/Conversion/LLVMCommon/TypeConverter.h"
+
+namespace quidditch {
+void populateDMAToLLVMConversionPatterns(mlir::ModuleOp moduleOp,
+                                         mlir::LLVMTypeConverter &converter,
+                                         mlir::RewritePatternSet &patterns);
+}
