+++ codegen/compiler/src/Quidditch/Dialect/DMA/CMakeLists.txt
@@ -1,0 +1,1 @@
+iree_add_all_subdirs()
