+++ codegen/compiler/src/Quidditch/Dialect/DMA/Extensions/CMakeLists.txt
@@ -1,0 +1,12 @@
+
+iree_cc_library(
+        NAME
+        DMACoreSpecializationOpInterfaceImpl
+        HDRS
+        "DMACoreSpecializationOpInterfaceImpl.h"
+        SRCS
+        "DMACoreSpecializationOpInterfaceImpl.cpp"
+        DEPS
+        Quidditch::Dialect::Snitch::IR::QuidditchSnitchDialect
+        Quidditch::Dialect::DMA::IR::DMADialect
+)
