+++ codegen/compiler/src/Quidditch/Dialect/DMA/Extensions/DMACoreSpecializationOpInterfaceImpl.h
@@ -1,0 +1,10 @@
+
+#pragma once
+
+namespace mlir {
+class DialectRegistry;
+}
+
+namespace quidditch::dma {
+void registerDMACoreSpecializationOpInterface(mlir::DialectRegistry &registry);
+}
