+++ codegen/compiler/src/Quidditch/Dialect/DMA/IR/DMAAttrs.h
@@ -1,0 +1,7 @@
+
+#pragma once
+
+#include "mlir/IR/Attributes.h"
+
+#define GET_ATTRDEF_CLASSES
+#include "Quidditch/Dialect/DMA/IR/DMAAttrs.h.inc"
