+++ codegen/compiler/src/Quidditch/Dialect/DMA/IR/DMADialect.h
@@ -1,0 +1,7 @@
+
+#pragma once
+
+#include "mlir/IR/Dialect.h"
+#include "mlir/IR/Operation.h"
+
+#include "Quidditch/Dialect/DMA/IR/DMADialect.h.inc"
