+++ codegen/compiler/src/Quidditch/Dialect/DMA/IR/DMAOps.h
@@ -1,0 +1,14 @@
+
+#pragma once
+
+#include "mlir/Bytecode/BytecodeOpInterface.h"
+#include "mlir/Dialect/Bufferization/IR/BufferizableOpInterface.h"
+#include "mlir/IR/BuiltinAttributes.h"
+#include "mlir/IR/OpImplementation.h"
+#include "mlir/Interfaces/InferTypeOpInterface.h"
+#include "mlir/Interfaces/SideEffectInterfaces.h"
+
+#include "DMATypes.h"
+
+#define GET_OP_CLASSES
+#include "Quidditch/Dialect/DMA/IR/DMAOps.h.inc"
