+++ codegen/compiler/src/Quidditch/Dialect/DMA/IR/DMATypes.h
@@ -1,0 +1,7 @@
+
+#pragma once
+
+#include "mlir/IR/Types.h"
+
+#define GET_TYPEDEF_CLASSES
+#include "Quidditch/Dialect/DMA/IR/DMATypes.h.inc"
