+++ codegen/compiler/src/Quidditch/Dialect/Snitch/IR/QuidditchSnitchDialect.td
@@ -14,8 +14,8 @@ def QuidditchSnitch_Dialect : Dialect {
   );
 
   let useDefaultAttributePrinterParser = 1;
-  let useDefaultTypePrinterParser = 1;
-  let hasConstantMaterializer = 1;
+  let useDefaultTypePrinterParser = 0;
+  let hasConstantMaterializer = 0;
 }
 
 #endif
