+++ codegen/compiler/src/Quidditch/Dialect/Snitch/IR/QuidditchSnitchOps.td
@@ -174,114 +174,6 @@ def QuidditchSnitch_MicrokernelFenceOp : QuidditchSnitch_Op<"microkernel_fence",
   }];
 }
 
-def QuidditchSnitch_StartTensorCopyOp : QuidditchSnitch_Op<"start_tensor_copy",
-  [Pure, AllRanksMatch<["copy", "result"]>,
-   DeclareOpInterfaceMethods<BufferizableOpInterface,
-    ["resultBufferizesToMemoryWrite", "bufferizesToMemoryRead",
-     "bufferizesToMemoryWrite", "getAliasingValues", "getBufferType",
-      "bufferize", "bufferizesToAllocation"]>]> {
-
-  let description = [{
-    Operation starting a copy of a tensor to L1 memory space, optionally adding
-    padding and returning it as a new tensor.
-    The contained values of the resulting tensor is in an unspecified state.
-    See `wait_for_tensor_copy` to transform the tensor value into a state
-    equal to `$copy`.
-
-    The operation may optionally add padding at the end of each dimension of
-    the tensor. Zero is used as the padding value.
-    The dimensions of the result tensor are computed using
-    `dims(copy)[i] + high_pad[i]`.
-
-    This operation is a noop if `$copy` is already in L1, no padding is added,
-    and bufferization can elide the copy.
-  }];
-
-  let arguments = (ins AnyRankedTensor:$copy,
-                       Variadic<Index>:$high_pad,
-                       OptionalAttr<DenseI64ArrayAttr>:$static_high_pad,
-                       UnitAttr:$undef_padding
-  );
-
-  let results = (outs
-    AnyRankedTensor:$result,
-    QuidditchSnitch_DMATokenType:$token
-  );
-
-  let assemblyFormat = [{
-    $copy `to` `L1`
-    ( `pad` `with` (`undef` $undef_padding^) : (`zero`)? `to`
-      custom<DynamicIndexList>($high_pad, $static_high_pad)^)?
-    `:` type($copy) `->` type($result) attr-dict
-  }];
-
-  let builders = [
-    OpBuilder<(ins "mlir::Value":$copy), [{
-      build($_builder, $_state, copy.getType(),
-            $_builder.getType<DMATokenType>(), copy,
-            /*high_pad=*/mlir::ValueRange(), /*static_high_pad=*/nullptr);
-    }]>
-  ];
-
-  let hasVerifier = 1;
-
-  let extraClassDeclaration = [{
-  private:
-    std::optional<bool>
-    elidesAllocation(const mlir::bufferization::BufferizationOptions &options = {},
-                      llvm::SmallVector<mlir::Value> *invocationStack = nullptr);
-  public:
-
-    bool hasPadding() {
-      return static_cast<bool>(getStaticHighPadAttr());
-    }
-
-    llvm::SmallVector<mlir::OpFoldResult> getMixedHighPad();
-  }];
-
-  let hasFolder = 1;
-}
-
-def QuidditchSnitch_WaitForTensorCopyOp : QuidditchSnitch_Op<"wait_for_tensor_copy",
-  [AllTypesMatch<["transfer_tensor", "result"]>, Pure,
-   DeclareOpInterfaceMethods<BufferizableOpInterface,
-    ["bufferizesToMemoryRead", "bufferizesToMemoryWrite", "getAliasingValues",
-     "bufferize", "mustBufferizeInPlace", "isNotConflicting"]>]> {
-
-  let description = [{
-    Operation asserting that a previous `start_tensor_copy` operation has finished.
-    Unless `token` is the result of an `completed_token` operation,
-    `transfer_tensor` and `token` must at runtime be a token and tensor yielded
-    by a `start_tensor_copy` operation and `copy` the original tensor used in
-    `start_tensor_copy`.
-
-    Once this operation returns, the returned tensor's values are guaranteed
-    equal to the `copy` operand and in L1 memory.
-
-    Note: The additional `copy` operand is given as it is effectively read by
-    this operation.
-    This additionally guarantees that the bufferization frame work does not
-    perform a write to the underlying buffer of `copy` while the transfer is
-    in progress.
-  }];
-
-  let arguments = (ins
-    AnyRankedTensor:$transfer_tensor,
-    QuidditchSnitch_DMATokenType:$token,
-    AnyRankedTensor:$copy
-  );
-
-  let results = (outs
-    AnyRankedTensor:$result
-  );
-
-  let assemblyFormat = [{
-    `of` $copy `:` type($copy) `to` $transfer_tensor `using` $token `->` type($transfer_tensor) attr-dict
-  }];
-
-  let hasFolder = 1;
-}
-
 def FlatI8MemRef : ConfinedType<MemRefOf<[I8]>, [HasStaticShapePred,
   HasAnyRankOfPred<[1]>], "one-dimensional i8 MemRef of a static size">;
 
@@ -292,107 +184,6 @@ def QuidditchSnitch_L1MemoryViewOp : QuidditchSnitch_Op<"l1_memory_view",
   let assemblyFormat = [{
     `->` type($result) attr-dict
   }];
-}
-
-def QuidditchSnitch_StartDMATransferOp : QuidditchSnitch_Op<"start_dma_transfer",
-  [MemoryEffects<[MemWrite]>, SameOperandsElementType, SameOperandsShape,
-   QuidditchSnitch_DMACoreSpecializationOpInterface]> {
-
-  let description = [{
-    Operation performing a DMA transfer from one MemRef to another.
-    The shapes (including dynamic ones at runtime) of both MemRefs must be
-    identical with different strides and offsets allowed.
-
-    The DMA operation is likely (but not guaranteed) to run asynchronous and
-    its completion only guaranteed by executing the `wait_for_dma_transfers`
-    operation with the token returned by this operation or a later one.
-  }];
-
-  let arguments = (ins
-    Arg<Non0RankedMemRefOf<[AnyType]>, "source", [MemRead]>:$source,
-    Arg<Non0RankedMemRefOf<[AnyType]>, "destination", [MemWrite]>:$dest
-  );
-
-  let results = (outs QuidditchSnitch_DMATokenType:$token);
-
-  let assemblyFormat = [{
-    `from` $source `:` type($source) `to` $dest `:` type($dest) attr-dict
-  }];
-
-  let hasFolder = 1;
-
-  let extraClassDeclaration = [{
-    void replaceWithNoop(mlir::RewriterBase& rewriter);
-  }];
-}
-
-def QuidditchSnitch_StartZeroMemTransferOp : QuidditchSnitch_Op<"start_zero_mem_transfer",
-  [MemoryEffects<[MemWrite]>,
-   QuidditchSnitch_DMACoreSpecializationOpInterface]> {
-
-  let description = [{
-
-  }];
-
-  let arguments = (ins
-    Arg<Non0RankedMemRefOf<[AnyType]>, "zeroed buffer", [MemWrite]>:$filled
-  );
-
-  let results = (outs QuidditchSnitch_DMATokenType:$token);
-
-  let assemblyFormat = [{
-    $filled `:` type($filled) attr-dict
-  }];
-
-  let extraClassDeclaration = [{
-    void replaceWithNoop(mlir::RewriterBase& rewriter);
-  }];
-}
-
-def QuidditchSnitch_WaitForDMATransfersOp
-  : QuidditchSnitch_Op<"wait_for_dma_transfers", [
-    QuidditchSnitch_DMACoreSpecializationOpInterface
-  ]> {
-
-  let description = [{
-    Operation awaiting for DMA transfers denoted by its tokens to be finished.
-  }];
-
-  let arguments = (ins
-    Variadic<QuidditchSnitch_DMATokenType>:$tokens
-  );
-
-  let assemblyFormat = [{
-    ($tokens^ `:` type($tokens))? attr-dict
-  }];
-
-  let hasFolder = 1;
-  let hasCanonicalizeMethod = 1;
-
-  let extraClassDeclaration = [{
-    bool needsSynchronization() {
-      return true;
-    }
-
-    void replaceWithNoop(mlir::RewriterBase& rewriter);
-  }];
-}
-
-def QuidditchSnitch_CompletedTokenOp
-  : QuidditchSnitch_Op<"completed_token", [Pure, ConstantLike]> {
-
-  let description = [{
-    Op returning a special value representing a completed DMA transfer.
-    Passing this token to `wait_for_dma_transfers` will always return immediately.
-  }];
-
-  let results = (outs QuidditchSnitch_DMATokenType:$token);
-
-  let assemblyFormat = [{
-    attr-dict
-  }];
-
-  let hasFolder = 1;
 }
 
 def QuidditchSnitch_BarrierOp : QuidditchSnitch_Op<"barrier"> {
