+++ codegen/compiler/src/Quidditch/Dialect/Snitch/IR/QuidditchSnitchTypes.td
@@ -7,12 +7,4 @@ include "mlir/IR/AttrTypeBase.td"
 class QuidditchSnitch_Type<string name, list<Trait> traits = []> :
   TypeDef<QuidditchSnitch_Dialect, name, traits>;
 
-def QuidditchSnitch_DMATokenType : QuidditchSnitch_Type<"DMAToken"> {
-  let mnemonic = "dma_token";
-
-  let description = [{
-    Type representing a potentially active DMA transfer.
-  }];
-}
-
 #endif
