+++ codegen/tests/Conversion/ConvertDMAToLLVM/dma_wait.mlir
@@ -2,7 +2,7 @@
 
 // CHECK-LABEL: @test
 // CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-func.func private @test(%arg0 : !quidditch_snitch.dma_token) {
+func.func private @test(%arg0 : !dma.token) {
   // CHECK: llvm.br ^[[BODY:[[:alnum:]]+]]
   // CHECK: ^[[BODY]]:
   // CHECK-NEXT: %[[ID:.*]] = llvm.inline_asm has_side_effects ".insn r 0x2b, 0, 0b100, $0, zero, zero
@@ -11,7 +11,7 @@ func.func private @test(%arg0 : !quidditch_snitch.dma_token) {
   // CHECK: %[[COND:.*]] = llvm.icmp "ult" %[[ID]], %[[ARG0]]
   // CHECK: llvm.cond_br %[[COND]], ^[[BODY]], ^[[CONT:[[:alnum:]]+]]
   // CHECK: ^[[CONT]]:
-  quidditch_snitch.wait_for_dma_transfers %arg0 : !quidditch_snitch.dma_token
+  dma.wait_for_transfers %arg0 : !dma.token
   // CHECK-NEXT: llvm.return
   return
 }
