+++ codegen/tests/Dialect/Snitch/IR/canonicalization.mlir
@@ -36,22 +36,6 @@ func.func @identical_argument(%arg0 : i32) {
   return
 }
 
-// CHECK-LABEL: @wait_gets_removed
-func.func @wait_gets_removed() {
-  // CHECK-NEXT: return
-  %0 = quidditch_snitch.completed_token
-  quidditch_snitch.wait_for_dma_transfers %0 : !quidditch_snitch.dma_token
-  return
-}
-
-// CHECK-LABEL: @noop_transfer
-func.func @noop_transfer(%arg0 : memref<?xf32>) -> !quidditch_snitch.dma_token {
-  // CHECK-NEXT: %[[R:.*]] = quidditch_snitch.completed_token
-  // CHECK-NEXT: return %[[R]]
-  %0 = quidditch_snitch.start_dma_transfer from %arg0 : memref<?xf32> to %arg0 : memref<?xf32>
-  return %0 : !quidditch_snitch.dma_token
-}
-
 // CHECK-LABEL: @pipeline_dead_block_arg(
 func.func @pipeline_dead_block_arg(%tensor : tensor<?xf32>) {
   %c0 = arith.constant 0 : index
@@ -90,98 +74,4 @@ func.func @pipeline_invariant(%tensor : tensor<?xf32>) {
     quidditch_snitch.pipeline_yield
   }
   return
-}
-
-// CHECK-LABEL: @tensor_wait_gets_removed
-// CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-// CHECK-SAME: %[[ARG1:[[:alnum:]]+]]
-func.func @tensor_wait_gets_removed(%arg0 : tensor<?xf32>, %arg1 : tensor<?xf32>) -> tensor<?xf32> {
-  // CHECK-NEXT: return %[[ARG1]]
-  %t = quidditch_snitch.completed_token
-  %0 = quidditch_snitch.wait_for_tensor_copy of %arg0 : tensor<?xf32> to %arg1 using %t -> tensor<?xf32>
-  return %0 : tensor<?xf32>
-}
-
-// CHECK-LABEL: @tensor_noop_transfer
-// CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-func.func @tensor_noop_transfer(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  // CHECK: %[[T2:.*]] = quidditch_snitch.completed_token
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy %[[ARG0]]
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 : tensor<?xf32> -> tensor<?xf32>
-  // CHECK: %[[R2:.*]] = quidditch_snitch.wait_for_tensor_copy of %[[ARG0]]
-  // CHECK-SAME: to %[[R]] using %[[T]]
-  %0 = quidditch_snitch.wait_for_tensor_copy of %arg0 : tensor<?xf32> to %r using %t -> tensor<?xf32>
-
-  // CHECK-NOT: wait_for_tensor_copy
-  %r2, %t2 = quidditch_snitch.start_tensor_copy %0 to L1 : tensor<?xf32> -> tensor<?xf32>
-
-  // CHECK: return %[[R2]], %[[T2]]
-  return %r2, %t2 : tensor<?xf32>, !quidditch_snitch.dma_token
-}
-
-// CHECK-LABEL: @tensor_noop_pad
-func.func @tensor_noop_pad(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy
-  // CHECK-NOT: pad with
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 pad with zero to [0] : tensor<?xf32> -> tensor<?xf32>
-  // CHECK-NEXT: return %[[R]], %[[T]]
-  return %r, %t : tensor<?xf32>, !quidditch_snitch.dma_token
-}
-
-// CHECK-LABEL: @tensor_pad_constant
-func.func @tensor_pad_constant(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  %zero = arith.constant 0 : index
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy
-  // CHECK-NOT: pad with
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 pad with zero to [%zero] : tensor<?xf32> -> tensor<?xf32>
-  // CHECK-NEXT: return %[[R]], %[[T]]
-  return %r, %t : tensor<?xf32>, !quidditch_snitch.dma_token
-}
-
-// CHECK-LABEL: @tensor_noop_transfer_pad
-// CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-func.func @tensor_noop_transfer_pad(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  // CHECK: %[[T2:.*]] = quidditch_snitch.completed_token
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy %[[ARG0]]
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 pad with zero to [1] : tensor<?xf32> -> tensor<?xf32>
-  // CHECK: %[[R2:.*]] = quidditch_snitch.wait_for_tensor_copy of %[[ARG0]]
-  // CHECK-SAME: to %[[R]] using %[[T]]
-  %0 = quidditch_snitch.wait_for_tensor_copy of %arg0 : tensor<?xf32> to %r using %t -> tensor<?xf32>
-
-  // CHECK-NOT: wait_for_tensor_copy
-  %r2, %t2 = quidditch_snitch.start_tensor_copy %0 to L1 : tensor<?xf32> -> tensor<?xf32>
-
-  // CHECK: return %[[R2]], %[[T2]]
-  return %r2, %t2 : tensor<?xf32>, !quidditch_snitch.dma_token
-}
-
-// CHECK-LABEL: @tensor_noop_transfer_pad_neg
-// CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-func.func @tensor_noop_transfer_pad_neg(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  // CHECK: start_tensor_copy
-  // CHECK: wait_for_tensor_copy
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy
-  // CHECK: return %[[R]], %[[T]]
-
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 : tensor<?xf32> -> tensor<?xf32>
-  %0 = quidditch_snitch.wait_for_tensor_copy of %arg0 : tensor<?xf32> to %r using %t -> tensor<?xf32>
-  %r2, %t2 = quidditch_snitch.start_tensor_copy %0 to L1 pad with zero to [1] : tensor<?xf32> -> tensor<?xf32>
-  return %r2, %t2 : tensor<?xf32>, !quidditch_snitch.dma_token
-}
-
-// CHECK-LABEL: @tensor_noop_transfer_same_padding
-// CHECK-SAME: %[[ARG0:[[:alnum:]]+]]
-func.func @tensor_noop_transfer_same_padding(%arg0 : tensor<?xf32>) -> (tensor<?xf32>, !quidditch_snitch.dma_token) {
-  // CHECK: %[[T2:.*]] = quidditch_snitch.completed_token
-  // CHECK: %[[R:.*]], %[[T:.*]] = quidditch_snitch.start_tensor_copy %[[ARG0]]
-  %r, %t = quidditch_snitch.start_tensor_copy %arg0 to L1 pad with zero to [1] : tensor<?xf32> -> tensor<?xf32>
-  // CHECK: %[[R2:.*]] = quidditch_snitch.wait_for_tensor_copy of %[[ARG0]]
-  // CHECK-SAME: to %[[R]] using %[[T]]
-  %0 = quidditch_snitch.wait_for_tensor_copy of %arg0 : tensor<?xf32> to %r using %t -> tensor<?xf32>
-
-  // CHECK-NOT: wait_for_tensor_copy
-  %r2, %t2 = quidditch_snitch.start_tensor_copy %0 to L1 pad with zero to [1] : tensor<?xf32> -> tensor<?xf32>
-
-  // CHECK: return %[[R2]], %[[T2]]
-  return %r2, %t2 : tensor<?xf32>, !quidditch_snitch.dma_token
 }
