+++ codegen/tools/quidditch-opt.cpp
@@ -2,6 +2,8 @@
 #include <mlir/Tools/mlir-opt/MlirOptMain.h>
 
 #include "Quidditch/Conversion/Passes.h"
+#include "Quidditch/Dialect/DMA/Extensions/DMACoreSpecializationOpInterfaceImpl.h"
+#include "Quidditch/Dialect/DMA/IR/DMADialect.h"
 #include "Quidditch/Dialect/Snitch/IR/QuidditchSnitchDialect.h"
 #include "Quidditch/Dialect/Snitch/Transforms/Passes.h"
 #include "Quidditch/Target/Passes.h"
@@ -26,8 +28,10 @@ int main(int argc, char **argv) {
 
   // Be lazy and support all upstream dialects as input dialects.
   DialectRegistry registry;
+  quidditch::dma::registerDMACoreSpecializationOpInterface(registry);
   iree_compiler::registerAllDialects(registry);
-  registry.insert<quidditch::Snitch::QuidditchSnitchDialect>();
+  registry.insert<quidditch::Snitch::QuidditchSnitchDialect,
+                  quidditch::dma::DMADialect>();
 
   quidditch::registerPasses();
   quidditch::registerConversionPasses();
